Finished project 3 code
I believe I have the code done. The mutexes were being used
incorrectly, I was assuming the test mutex function wouldn't set
them, but it does. Removing the while loops fixed this issue. I
still don't know if the code is fully working, but it seems to be.
I'll know for sure tomorrow when I analyze the data.